+++ src/editor.rs
@@ -9,7 +9,9 @@ use ratatui::{
 use crate::app_state::AppState;
 
 pub fn render_editor(frame: &mut Frame, area: Rect, app_state: &mut AppState) {
-    app_state.ui_state.set_editor_offset(area.x, area.y);
+    app_state
+        .ui_state
+        .set_editor_offset(area.x as usize, area.y as usize);
 
     let lines_number = app_state.ui_state.lines.len();
     let text: Vec<Line> = app_state
